requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0